Clean up Queue Companion Objects
